+++ packages/backend/src/libp2p/connectionsManager.ts
@@ -1,16 +1,22 @@
 import { Crypto } from '@peculiar/webcrypto'
 import { Agent } from 'https'
-import { HttpsProxyAgent } from 'https-proxy-agent'
-import type { Libp2p, createLibp2p as l2l } from 'libp2p'
-import type { bootstrap as bootstrapType } from '@libp2p/bootstrap'
-import type { kadDHT as kadDHTType } from '@libp2p/kad-dht'
+import createHttpsProxyAgent from 'https-proxy-agent'
+
+import { peerIdFromKeys } from '@libp2p/peer-id'
+import { createLibp2p, Libp2p } from 'libp2p'
+import { noise } from '@chainsafe/libp2p-noise'
+import { gossipsub } from '@chainsafe/libp2p-gossipsub'
+import { mplex } from '@libp2p/mplex'
+import { bootstrap } from '@libp2p/bootstrap'
+import { kadDHT } from '@libp2p/kad-dht'
+import { createServer } from 'it-ws'
 
 import { webSockets } from './websocketOverTor/index'
 import { all } from './websocketOverTor/filters'
 
 import { DateTime } from 'luxon'
 
-import SocketIO from 'socket.io'
+import type SocketIO from 'socket.io'
 import * as os from 'os'
 import path from 'path'
 import fs from 'fs'
@@ -51,11 +57,7 @@ import {
 } from '@quiet/state-manager'
 
 import { ConnectionsManagerOptions } from '../common/types'
-import {
-  createLibp2pAddress,
-  createLibp2pListenAddress,
-  getPorts
-} from '../common/utils'
+
 import { QUIET_DIR_PATH } from '../constants'
 import { Storage } from '../storage'
 import { Tor } from '../torManager'
@@ -67,7 +69,8 @@ import { RegistrationEvents } from '../registration/types'
 import { StorageEvents } from '../storage/types'
 import { Libp2pEvents } from './types'
 import PeerId, { JSONPeerId } from 'peer-id'
-import { importDynamically } from './utils'
+
+import { createLibp2pAddress, createLibp2pListenAddress, getPorts } from '../common/utils'
 
 const log = logger('conn')
 interface InitStorageParams {
@@ -109,35 +112,6 @@ export interface InitLibp2pParams {
   bootstrapMultiaddrs: string[]
   certs: Certificates
 }
-
-let peerIdFromKeys = null
-let createLibp2p = null
-let noise = null
-let gossipsub = null
-let mplex = null
-let bootstrap = null
-let kadDHT = null
-let createServer = null
-
-void (async () => {
-  const { peerIdFromKeys: peerIdFromKeysImported } = await importDynamically('@libp2p/peer-id/dist/src/index.js')
-  const { createLibp2p: createLibp2pImported }: { createLibp2p: typeof l2l } = await importDynamically('libp2p/dist/src/index.js')
-  const { noise: noiseImported } = await importDynamically('@chainsafe/libp2p-noise/dist/src/index.js')
-  const { gossipsub: gossipsubImported } = await importDynamically('@chainsafe/libp2p-gossipsub/dist/src/index.js')
-  const { mplex: mplexImported } = await importDynamically('@libp2p/mplex/dist/src/index.js')
-  const { bootstrap: bootstrapImported }: { bootstrap: typeof bootstrapType } = await importDynamically('@libp2p/bootstrap/dist/src/index.js')
-  const { kadDHT: kadDHTImported }: { kadDHT: typeof kadDHTType } = await importDynamically('@libp2p/kad-dht/dist/src/index.js')
-  const { createServer: createServerImported } = await importDynamically('it-ws/dist/src/server.js')
-
-  peerIdFromKeys = peerIdFromKeysImported
-  createLibp2p = createLibp2pImported
-  noise = noiseImported
-  gossipsub = gossipsubImported
-  mplex = mplexImported
-  bootstrap = bootstrapImported
-  kadDHT = kadDHTImported
-  createServer = createServerImported
-})()
 
 export class ConnectionsManager extends EventEmitter {
   registration: CertificateRegistration
@@ -208,7 +182,7 @@ export class ConnectionsManager extends EventEmitter {
 
     log(`Creating https proxy agent: ${this.httpTunnelPort}`)
 
-    return new HttpsProxyAgent({ port: this.httpTunnelPort, host: 'localhost' })
+    return createHttpsProxyAgent({ port: this.httpTunnelPort, host: 'localhost' })
   }
 
   public init = async () => {