+++ packages/backend/src/registrarTest/torRegistrarTest.ts
@@ -1,8 +1,7 @@
 import express from 'express'
-import { HttpsProxyAgent } from 'https-proxy-agent'
+import createHttpsProxyAgent from 'https-proxy-agent'
 import fetch, { Response } from 'node-fetch'
 import fs from 'fs'
-import { getPorts } from '../common/utils'
 import {
   createTmpDir,
   spawnTorProcess,
@@ -13,6 +12,7 @@ import logger from '../logger'
 import { EventEmitter } from 'events'
 import { Command } from 'commander'
 import { Tor, GetInfoTorSignal } from '../torManager/torManager'
+import { getPorts } from '../common/utils'
 const program = new Command()
 
 enum TestMode {
@@ -98,7 +98,7 @@ const createServer = async (port, serverAddress: string) => {
 }
 
 const createAgent = (httpTunnelPort: number) => {
-  return new HttpsProxyAgent({ port: httpTunnelPort, host: 'localhost' })
+  return createHttpsProxyAgent({ port: httpTunnelPort, host: 'localhost' })
 }
 
 const sendRequest = async (