+++ packages/state-manager/CHANGELOG.md
@@ -3,6 +3,142 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.19.19-alpha.14](https://github.com/TryQuiet/quiet/compare/@quiet/state-manager@0.19.19-alpha.13...@quiet/state-manager@0.19.19-alpha.14) (2023-02-08)
+
+**Note:** Version bump only for package @quiet/state-manager
+
+
+
+
+
+## [0.19.19-alpha.13](https://github.com/TryQuiet/quiet/compare/@quiet/state-manager@0.19.19-alpha.12...@quiet/state-manager@0.19.19-alpha.13) (2023-02-08)
+
+**Note:** Version bump only for package @quiet/state-manager
+
+
+
+
+
+## [0.19.19-alpha.12](https://github.com/TryQuiet/quiet/compare/@quiet/state-manager@0.19.19-alpha.5...@quiet/state-manager@0.19.19-alpha.12) (2023-02-07)
+
+**Note:** Version bump only for package @quiet/state-manager
+
+
+
+
+
+## [0.19.19-alpha.5](https://github.com/TryQuiet/quiet/compare/@quiet/state-manager@0.19.19-alpha.4...@quiet/state-manager@0.19.19-alpha.5) (2023-02-07)
+
+**Note:** Version bump only for package @quiet/state-manager
+
+
+
+
+
+## [0.19.19-alpha.4](https://github.com/TryQuiet/quiet/compare/@quiet/state-manager@0.19.19-alpha.3...@quiet/state-manager@0.19.19-alpha.4) (2023-02-06)
+
+**Note:** Version bump only for package @quiet/state-manager
+
+
+
+
+
+## [0.19.19-alpha.3](https://github.com/TryQuiet/quiet/compare/@quiet/state-manager@0.19.19-alpha.2...@quiet/state-manager@0.19.19-alpha.3) (2023-02-06)
+
+**Note:** Version bump only for package @quiet/state-manager
+
+
+
+
+
+## [0.19.19-alpha.2](https://github.com/TryQuiet/quiet/compare/@quiet/state-manager@0.19.7-alpha.0...@quiet/state-manager@0.19.19-alpha.2) (2023-02-06)
+
+**Note:** Version bump only for package @quiet/state-manager
+
+
+
+
+
+## [0.19.7-alpha.0](https://github.com/TryQuiet/quiet/compare/@quiet/state-manager@0.19.6-alpha.0...@quiet/state-manager@0.19.7-alpha.0) (2023-02-02)
+
+**Note:** Version bump only for package @quiet/state-manager
+
+
+
+
+
+## [0.19.6-alpha.0](https://github.com/TryQuiet/quiet/compare/@quiet/state-manager@0.19.5-alpha.0...@quiet/state-manager@0.19.6-alpha.0) (2023-02-02)
+
+**Note:** Version bump only for package @quiet/state-manager
+
+
+
+
+
+## [0.19.5-alpha.0](https://github.com/TryQuiet/quiet/compare/@quiet/state-manager@0.19.4-alpha.0...@quiet/state-manager@0.19.5-alpha.0) (2023-02-01)
+
+**Note:** Version bump only for package @quiet/state-manager
+
+
+
+
+
+## [0.19.4-alpha.0](https://github.com/TryQuiet/quiet/compare/@quiet/state-manager@0.19.3-alpha.0...@quiet/state-manager@0.19.4-alpha.0) (2023-02-01)
+
+**Note:** Version bump only for package @quiet/state-manager
+
+
+
+
+
+## [0.19.3-alpha.0](https://github.com/TryQuiet/quiet/compare/@quiet/state-manager@0.17.4-alpha.0...@quiet/state-manager@0.19.3-alpha.0) (2023-01-31)
+
+**Note:** Version bump only for package @quiet/state-manager
+
+
+
+
+
+## [0.19.2-alpha.0](https://github.com/TryQuiet/quiet/compare/@quiet/state-manager@0.17.4-alpha.0...@quiet/state-manager@0.19.2-alpha.0) (2023-01-30)
+
+**Note:** Version bump only for package @quiet/state-manager
+
+
+
+
+
+## [0.19.1-alpha.0](https://github.com/TryQuiet/quiet/compare/@quiet/state-manager@0.17.4-alpha.0...@quiet/state-manager@0.19.1-alpha.0) (2023-01-30)
+
+**Note:** Version bump only for package @quiet/state-manager
+
+
+
+
+
+# [0.19.0-alpha.2](https://github.com/TryQuiet/quiet/compare/@quiet/state-manager@0.17.4-alpha.0...@quiet/state-manager@0.19.0-alpha.2) (2023-01-27)
+
+**Note:** Version bump only for package @quiet/state-manager
+
+
+
+
+
+# [0.19.0-alpha.1](https://github.com/ZbayApp/monorepo/compare/@quiet/state-manager@0.17.4-alpha.0...@quiet/state-manager@0.19.0-alpha.1) (2023-01-27)
+
+**Note:** Version bump only for package @quiet/state-manager
+
+
+
+
+
+# [0.19.0-alpha.0](https://github.com/TryQuiet/quiet/compare/@quiet/state-manager@0.17.4-alpha.0...@quiet/state-manager@0.19.0-alpha.0) (2023-01-27)
+
+**Note:** Version bump only for package @quiet/state-manager
+
+
+
+
+
 ## [0.17.4-alpha.0](https://github.com/TryQuiet/quiet/compare/@quiet/state-manager@0.17.3-alpha.0...@quiet/state-manager@0.17.4-alpha.0) (2023-01-15)
 
 **Note:** Version bump only for package @quiet/state-manager